Move problem report code to it's own file
